Add Keypad-Ext
This was a temporary bodge board for the Revision 2 hardware, which added
a PCA9536 I/O expander to allow LED12 to be driven separately.  Also
the presence of PCA9536 on the I2C bus was also an indicator we were
running on "revision 2" hardware, which had some of the capacitive
sense lines moved around compare to revision 1 hardware.

In the end I didn't end up using these files; I redrew the schematic
using EasyEDA in order to try out JLCPCB's PCBA service for the first
time.  But the schematic here is the same as the one I redrew and had
fabricated.